speech_to_text/requirements.txt
google-cloud-storage==1.30.0
google-cloud-speech==2.0.0